COMP: Fix -Wnarrowing warning in vtkPlanarContourToClosedSurfaceConversionRule
This commit fixes the following warning:

/path/to/SlicerRT/DicomRtImportExport/ConversionRules/vtkPlanarContourToClosedSurfaceConversionRule.cxx:1342:34: warning: narrowing conversion of ‘ceil(((bounds[1] - bounds[0]) / spacing[0]))’ from ‘double’ to ‘int’ inside { } [-Wnarrowing]
   int dimensions[3] = { std::ceil(static_cast<double>((bounds[1]-bounds[0])/spacing[0])),
                                  ^
/path/to/SlicerRT/DicomRtImportExport/ConversionRules/vtkPlanarContourToClosedSurfaceConversionRule.cxx:1343:34: warning: narrowing conversion of ‘ceil(((bounds[3] - bounds[2]) / spacing[1]))’ from ‘double’ to ‘int’ inside { } [-Wnarrowing]
                         std::ceil(static_cast<double>((bounds[3]-bounds[2])/spacing[1])),
                                  ^

+++ DicomRtImportExport/ConversionRules/vtkPlanarContourToClosedSurfaceConversionRule.cxx
@@ -1339,8 +1339,8 @@ void vtkPlanarContourToClosedSurfaceConversionRule::CreateEndCapContour(vtkPolyD
   bounds[2] -= (this->ImagePadding[0]/2) * spacing[1];
   bounds[3] += (this->ImagePadding[0]/2) * spacing[1];
 
-  int dimensions[3] = { std::ceil((bounds[1]-bounds[0])/spacing[0]),
-                        std::ceil((bounds[3]-bounds[2])/spacing[1]),
+  int dimensions[3] = { static_cast<int>(std::ceil((bounds[1]-bounds[0])/spacing[0])),
+                        static_cast<int>(std::ceil((bounds[3]-bounds[2])/spacing[1])),
                         1 };
   double origin[3] = {bounds[0], bounds[2], bounds[4]};
   int extent[6] = {0, dimensions[0]-1, 0, dimensions[1]-1, 0, 0};
